chore: Configure do not trim trailing whitespace
The Hegel codebase has trailing whitespace in various files
and documentation.

By text editor will remove the trailing whitespace by default
on lines which adds noise to the git diff due to whitespace
removal.

I've enabled the settings to keep trailing whitespace upon
save in the text editor and i have added this setting to git so
that other contributors do not run into this issue.